optimized vtables in windows part hopefully fixed crash #3666 onSecond and onMinute ticks supports 64bit times first try to fix Hub reconnect (THX adrian_007 for some ideas)

+++ trunk/client/ClientManager.cpp
@@ -491,7 +491,7 @@ void ClientManager::on(Load, SimpleXML&) throw() {
 	users.insert(make_pair(getMe()->getCID(), getMe()));
 }
 
-void ClientManager::on(TimerManagerListener::Minute, uint32_t /*aTick*/) throw() {
+void ClientManager::on(TimerManagerListener::Minute, uint64_t /*aTick*/) throw() {
 	Lock l(cs);
 
 	// Collect some garbage...
