+++ trunk/client/HashManager.h
@@ -219,7 +219,7 @@ private:
 		Lock l(cs);
 		store.rebuild();
 	}
-	void on(TimerManagerListener::Minute, uint32_t) throw() {
+	void on(TimerManagerListener::Minute, uint64_t) throw() {
 		Lock l(cs);
 		store.save();
 	}
