+++ trunk/client/ShareManager.h
@@ -304,7 +304,7 @@ private:
 	}
 	
 	// TimerManagerListener
-	void on(TimerManagerListener::Minute, uint32_t tick) throw();
+	void on(TimerManagerListener::Minute, uint64_t tick) throw();
 	void load(SimpleXML& aXml);
 	void save(SimpleXML& aXml);
