+++ trunk/client/TimerManager.h
@@ -41,8 +41,8 @@ public:
 	typedef X<1> Minute;
 
 	// We expect everyone to implement this...
-	virtual void on(Second, uint32_t) throw() { }
-	virtual void on(Minute, uint32_t) throw() { }
+	virtual void on(Second, uint64_t) throw() { }
+	virtual void on(Minute, uint64_t) throw() { }
 };
 
 class TimerManager : public Speaker<TimerManagerListener>, public Singleton<TimerManager>, public Thread
